+++ packages/dependy_flutter/example/example_9/services/module.dart
@@ -1,0 +1,21 @@
+import 'package:dependy/dependy.dart';
+
+import 'counter_service.dart';
+import 'logger_service.dart';
+
+final example9ServicesModule = DependyModule(
+  providers: {
+    DependyProvider<LoggerService>(
+      (_) => ConsoleLoggerService(),
+    ),
+    DependyProvider<CounterService>(
+      (dependy) async {
+        final loggerService = await dependy<LoggerService>();
+        return CounterServiceImpl(1, loggerService);
+      },
+      dependsOn: {
+        LoggerService,
+      },
+    ),
+  },
+);