+++ packages/dependy_flutter/lib/dependy_flutter.dart
@@ -1,4 +1,5 @@
-library dependy_flutter;
+library;
 
+export 'src/async_builder.dart';
 export 'src/impl.dart';
 export 'src/watch.dart';